Refactor statements again and try to process comments
