+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/CommentStmt.java
@@ -19,7 +19,7 @@ public class CommentStmt extends Stmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         return getContent();
     }
 }
