+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/CommonStmt.java
@@ -18,7 +18,7 @@ public class CommonStmt extends SpecificationStmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         var blocks = getBlocks();
         var blocksCode =  blocks.stream()
                 .map(CommonBlock::getCode)
