+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/ContainsStmt.java
@@ -5,13 +5,13 @@ import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
 
 import java.util.Collection;
 
-public class ContainsStmt extends Stmt {
+public class ContainsStmt extends ExecutableStmt {
     public ContainsStmt(DataStore data, Collection<? extends FortranNode> children) {
         super(data, children);
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         return "contains";
     }
 }
