+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/ParameterStmt.java
@@ -19,7 +19,7 @@ public class ParameterStmt extends Stmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         var defsCode = getDefs().stream()
                 .map(NamedConstantDef::getCode)
                 .collect(Collectors.joining(", "));
