+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/UseStmt.java
@@ -21,7 +21,7 @@ public class UseStmt extends Stmt {
     }
 
     @Override
-    public String getCode() {
-        return FortranKeyword.USE.toString() + " " + getName();
+    public String getStmtCode() {
+        return FortranKeyword.USE + " " + getName();
     }
 }
