+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/datastmt/DataStmt.java
@@ -19,7 +19,7 @@ public class DataStmt extends DeclarationStmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         var dataSets = getDataSets();
 
         var dataSetsCode = dataSets.stream()
