+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/ifstmt/ElseIfStmt.java
@@ -4,13 +4,14 @@ import org.suikasoft.jOptions.Interfaces.DataStore;
 import pt.up.fe.specs.fortran.ast.FortranKeyword;
 import pt.up.fe.specs.fortran.ast.nodes.FortranNode;
 import pt.up.fe.specs.fortran.ast.nodes.expr.Expr;
+import pt.up.fe.specs.fortran.ast.nodes.stmt.Stmt;
 
 import java.util.Collection;
 
 /**
  * R1136 else-if-stmt
  */
-public class ElseIfStmt extends FortranNode {
+public class ElseIfStmt extends Stmt {
     public ElseIfStmt(DataStore data, Collection<? extends FortranNode> children) {
         super(data, children);
     }
@@ -19,7 +20,7 @@ public class ElseIfStmt extends FortranNode {
         return getChild(Expr.class, 0);
     }
 
-    public String getCode() {
+    public String getStmtCode() {
         var condition = getCondition();
         var nameOpt = ((IfConstruct) getParent().getParent()).getName();
