+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/selectcase/DefaultCaseStmt.java
@@ -12,7 +12,7 @@ public class DefaultCaseStmt extends CaseStmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         var nameOpt = getAncestor(CaseConstruct.class).getName();
 
         var code = new StringBuilder();
