+++ FortranAst/src/pt/up/fe/specs/fortran/ast/nodes/stmt/selectcase/SelectCaseStmt.java
@@ -18,7 +18,7 @@ public class SelectCaseStmt extends Stmt {
     }
 
     @Override
-    public String getCode() {
+    public String getStmtCode() {
         var nameOpt = getAncestor(CaseConstruct.class).getName();
         var expr = getExpr();
